api/requirements.txt
fastapi==0.100.0
tortoise-orm==0.19.3
tortoise-orm[asyncpg]==0.19.3
uvicorn==0.22.0
pydantic==2.0.2
aerich==0.7.1
